+++ src/components/Counter/Counter.styles.ts
@@ -11,21 +11,18 @@ import {
   getSizeMediumX,
   getSpacingMicro,
   getSpacingSmall,
-  getTypographyStyles,
 } from '../../common/themeSelectors';
 import { CounterSizes } from './Counter.types';
 
-export const Label = styled.Text<{ theme: Theme }>(({ theme }): CSSObject => {
-  const { subtitle2 } = getTypographyStyles(theme);
-
-  return ({
-    color: getColorMediumEmphasis(theme),
-    fontSize: subtitle2.fontSize,
-    fontWeight: subtitle2.fontWeight,
-    letterSpacing: subtitle2.letterSpacing,
-    marginBottom: getSpacingMicro(theme),
-  });
-});
+export const Label = styled.Text<{ theme: Theme }>(({ theme }): CSSObject => ({
+  color: getColorMediumEmphasis(theme),
+  fontFamily: theme.counter.label.primary.fontFamily,
+  fontSize: theme.counter.label.fontSize,
+  fontWeight: theme.counter.label.primary.fontWeight,
+  letterSpacing: theme.counter.label.letterSpacing,
+  lineHeight: theme.counter.label.fontSize * theme.counter.label.lineHeight,
+  marginBottom: getSpacingMicro(theme),
+}));
 
 export const Container = styled.View<{
   size: CounterSizes, theme: Theme
@@ -40,20 +37,18 @@ export const Container = styled.View<{
   height: getSize(theme, size),
 }));
 
-export const Input = styled.TextInput<{ theme: Theme }>(({ theme }): CSSObject => {
-  const { body1 } = getTypographyStyles(theme);
-
-  return ({
-    borderColor: getColorLowEmphasis(theme),
-    borderLeftWidth: 1,
-    borderRightWidth: 1,
-    color: getColorHighEmphasis(theme),
-    fontSize: body1.fontSize,
-    fontWeight: body1.fontWeight,
-    height: '100%',
-    letterSpacing: body1.letterSpacing,
-    paddingHorizontal: getSpacingSmall(theme),
-    textAlign: 'center',
-    width: getSizeMediumX(theme),
-  });
-});
+export const Input = styled.TextInput<{ theme: Theme }>(({ theme }): CSSObject => ({
+  borderColor: getColorLowEmphasis(theme),
+  borderLeftWidth: 1,
+  borderRightWidth: 1,
+  color: getColorHighEmphasis(theme),
+  fontFamily: theme.counter.content.primary.fontFamily,
+  fontSize: theme.counter.content.fontSize,
+  fontWeight: theme.counter.content.primary.fontWeight,
+  height: '100%',
+  letterSpacing: theme.counter.content.letterSpacing,
+  lineHeight: theme.counter.content.fontSize * theme.counter.content.lineHeight,
+  paddingHorizontal: getSpacingSmall(theme),
+  textAlign: 'center',
+  width: getSizeMediumX(theme),
+}));